requirements.txt
flask
firebase-admin
python-dotenv
flask-apscheduler
reportlab
requests
langchain
langchain-google-genai
langchain-huggingface
langchain-community
faiss-cpu
